Fix single user install + auto bundle umu_version.json
A partial cause for the user-install to fail was the makefile was
attempting to copy umu_version.json in the system wide site-packages.
This commit fixes this issue by adding umu_version.json as an artifact in
pyproject.toml, and changing the makefile rules to have umu-dist depend
on version. This way, python build/install will handle this file for us.
This should also make things easier for packagers.

The installation part is handled using pipx instead of pip --user since
the later may fail on some distros (even with --user).

Now the single user install is as simple as:

  make user-install

unisntall:

  make user-uninstall

no need to use configure.sh.

TODO:
[] remove user-install option from configure.sh
[] update README.md
[] remove references to PYTHONDIR in Makefuke.in that cause issues with
packaging in general
[] address issue with the umu-launcher target, whould it be installed in
single user, and what command to use?